Lots of changes to support controlling and controlled candidates and to add compiling checklists
git-svn-id: http://svn.littleshoot.org/svn/littleshoot/trunk/common/ice@753 8d7aff80-242d-0410-bcee-ed6f15385b3d

+++ src/main/java/org/lastbamboo/common/ice/IceCandidateComparator.java
@@ -14,8 +14,8 @@ public class IceCandidateComparator implements Comparator<IceCandidate>
     public int compare(final IceCandidate candidate1, 
         final IceCandidate candidate2)
         {
-        final int priority1 = candidate1.getPriority();
-        final int priority2 = candidate2.getPriority();
+        final long priority1 = candidate1.getPriority();
+        final long priority2 = candidate2.getPriority();
         
         if (priority1 > priority2) return -1;
         if (priority1 < priority2) return 1;
